[refactor/#52]: 비밀번호 패턴 순서에 관계 없도록 변경 및 패키지 이동
