+++ src/main/java/com/woohaengshi/backend/dto/request/auth/SignInRequest.java
@@ -1,4 +1,4 @@
-package com.woohaengshi.backend.dto.request.studyrecord.auth;
+package com.woohaengshi.backend.dto.request.auth;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Pattern;
@@ -12,8 +12,7 @@ public class SignInRequest {
     @Email(message = "잘못된 이메일 형식입니다.")
     private String email;
 
-    @Length(min = 8, max = 20)
-    @Pattern(regexp = "^[a-zA-Z]+\\d+[!@#$%^&*]+$")
+    @Pattern(regexp = "^(?=.*[a-zA-Z])(?=.*\\d)(?=.*[!@#$%^&*]).{8,20}$", message = "비밀번호는 영어, 숫자, 특수문자를 포함해야합니다.")
     private String password;
 
     public SignInRequest(String email, String password) {
