+++ src/main/java/com/woohaengshi/backend/dto/request/auth/SignUpRequest.java
@@ -25,8 +25,7 @@ public class SignUpRequest {
     @Email(message = "잘못된 이메일 형식입니다.")
     private String email;
 
-    @Length(min = 8, max = 20, message = "비밀번호는 8자 이상 20자 이하여야 합니다. ")
-    @Pattern(regexp = "^[a-zA-Z]+\\d+[!@#$%^&*]+$", message = "비밀번호는 영어, 숫자, 특수문자를 포함해야합니다.")
+    @Pattern(regexp = "^(?=.*[a-zA-Z])(?=.*\\d)(?=.*[!@#$%^&*]).{8,20}$", message = "비밀번호는 영어, 숫자, 특수문자를 포함해야합니다.")
     private String password;
 
     public SignUpRequest(String name, Course course, String email, String password) {
