+++ src/main/java/com/woohaengshi/backend/service/auth/AuthService.java
@@ -1,7 +1,7 @@
 package com.woohaengshi.backend.service.auth;
 
 import com.woohaengshi.backend.dto.request.auth.SignUpRequest;
-import com.woohaengshi.backend.dto.request.studyrecord.auth.SignInRequest;
+import com.woohaengshi.backend.dto.request.auth.SignInRequest;
 import com.woohaengshi.backend.dto.result.SignInResult;
 
 import org.springframework.http.ResponseCookie;
